# Improve packfile reading performance