--- a/plumbing/format/packfile/common.go
+++ b/plumbing/format/packfile/common.go
@@ -2,11 +2,9 @@ package packfile
 
 import (
 	"bytes"
-	"errors"
 	"io"
 	"sync"
 
-	"gopkg.in/src-d/go-git.v4/plumbing"
 	"gopkg.in/src-d/go-git.v4/plumbing/storer"
 	"gopkg.in/src-d/go-git.v4/utils/ioutil"
 )
@@ -32,8 +30,12 @@ func UpdateObjectStorage(s storer.Storer, packfile io.Reader) error {
 		return WritePackfileToObjectStorage(pw, packfile)
 	}
 
-	updater := newPackfileStorageUpdater(s)
-	_, err := NewParser(NewScanner(packfile), updater).Parse()
+	p, err := NewParserWithStorage(NewScanner(packfile), s)
+	if err != nil {
+		return err
+	}
+
+	_, err = p.Parse()
 	return err
 }
 
@@ -57,57 +59,4 @@ var bufPool = sync.Pool{
 	New: func() interface{} {
 		return bytes.NewBuffer(nil)
 	},
-}
-
-var errMissingObjectContent = errors.New("missing object content")
-
-type packfileStorageUpdater struct {
-	storer.Storer
-	lastSize int64
-	lastType plumbing.ObjectType
-}
-
-func newPackfileStorageUpdater(s storer.Storer) *packfileStorageUpdater {
-	return &packfileStorageUpdater{Storer: s}
-}
-
-func (p *packfileStorageUpdater) OnHeader(count uint32) error {
-	return nil
-}
-
-func (p *packfileStorageUpdater) OnInflatedObjectHeader(
-	t plumbing.ObjectType,
-	objSize int64,
-	pos int64,
-) error {
-	if p.lastSize > 0 || p.lastType != plumbing.InvalidObject {
-		return errMissingObjectContent
-	}
-
-	p.lastType = t
-	p.lastSize = objSize
-	return nil
-}
-
-func (p *packfileStorageUpdater) OnInflatedObjectContent(
-	h plumbing.Hash,
-	pos int64,
-	crc uint32,
-	content []byte,
-) error {
-	obj := new(plumbing.MemoryObject)
-	obj.SetSize(p.lastSize)
-	obj.SetType(p.lastType)
-	if _, err := obj.Write(content); err != nil {
-		return err
-	}
-
-	_, err := p.SetEncodedObject(obj)
-	p.lastSize = 0
-	p.lastType = plumbing.InvalidObject
-	return err
-}
-
-func (p *packfileStorageUpdater) OnFooter(h plumbing.Hash) error {
-	return nil
 }