… .travis.yml
… common_test.go
… plumbing/format/gitignore/dir_test.go
… plumbing/format/idxfile/writer_test.go
… plumbing/format/packfile/decoder.go
… plumbing/format/packfile/decoder_test.go
… plumbing/format/packfile/encoder_advanced_test.go
… plumbing/format/packfile/encoder_test.go
… plumbing/format/packfile/packfile_test.go
… plumbing/format/packfile/parser.go
… plumbing/format/packfile/parser_test.go
… plumbing/object/blob_test.go
… plumbing/object/difftree_test.go
… plumbing/object/object_test.go
… plumbing/transport/test/receive_pack.go
… plumbing/transport/test/upload_pack.go
… repository_test.go
… storage/filesystem/dotgit/writers.go
… storage/filesystem/object.go
… storage/filesystem/object_test.go
… worktree.go
… worktree_test.go