@erizocosmico 2018-08-01
Tests fixed (windows as well)