@erizocosmico 2018-08-07
Just committed using the parser to populate storage and a ton of bug and test fixes.

The only things that are missing are:
- Make parser work with non-seekable readers
- reference delta not found errors in tests, which I really don't know where they're coming from