@erizocosmico 2018-08-08
- Parser can work now with non-seekable readers (providing a `storer.EncodedObjectStorer` a la old decoder)
- Fixed how ref deltas are resolved. Before, it didn't take into account that ref deltas may be referencing deltas that have not been resolved yet
- Some bug fixes
- Tests passing now

The only think I feel like is missing from this before reviewing it is remove a hacky workaround that I did on `Packfile`.

Right now, `DiskObject`s require the file to be opened until they are read. On the filesystem storage `DecodeObjectAt` we open the file, get the object, and then close it, which cannot be like that. (The workaround right now is to convert it to a MemoryObject before returning, but that defeats the purpose of having a DiskObject in the first place).
There's a few possible solutions for that:

- Give the DiskObject a way to auto open/close the file by itself instead of a Packfile
- Wrap DiskObject with something that also receives the file and closes it when it's read. Downside of this: can only read once.

WDYT? @ajnavarro