@erizocosmico 2018-08-08
Windows tests seem to be failing because of an access denied removing fixtures 🤷‍♀️