@erizocosmico 2018-08-09
Just added a benchmark for `Parser`.

Before, using `Decoder`:

```
go test -benchmem -run=^$ gopkg.in/src-d/go-git.v4/plumbing/format/packfile -bench ^BenchmarkDecode$

goos: darwin
goarch: amd64
pkg: gopkg.in/src-d/go-git.v4/plumbing/format/packfile
BenchmarkDecode/https://github.com/git-fixtures/root-references.git-4         	     500	   3168400 ns/op	 1106147 B/op	    1211 allocs/op
BenchmarkDecode/https://github.com/git-fixtures/basic.git-4                   	     500	   2575930 ns/op	 1062023 B/op	     612 allocs/op
BenchmarkDecode/https://github.com/git-fixtures/basic.git#01-4                	     500	   2612729 ns/op	 1052947 B/op	     583 allocs/op
BenchmarkDecode/https://github.com/git-fixtures/basic.git#02-4                	     500	   2377921 ns/op	 1054738 B/op	     529 allocs/op
BenchmarkDecode/https://github.com/src-d/go-git.git-4                         	       2	 500670936 ns/op	190135260 B/op	   65837 allocs/op
BenchmarkDecode/https://github.com/git-fixtures/tags.git-4                    	   10000	    164102 ns/op	  117410 B/op	     162 allocs/op
BenchmarkDecode/https://github.com/spinnaker/spinnaker.git-4                  	      10	 142586155 ns/op	64122908 B/op	  100138 allocs/op
BenchmarkDecode/https://github.com/jamesob/desk.git-4                         	     100	  18939530 ns/op	 6878197 B/op	   11440 allocs/op
BenchmarkDecode/https://github.com/cpcs499/Final_Pres_P.git-4                 	   30000	     58652 ns/op	   52102 B/op	      64 allocs/op
BenchmarkDecode/https://github.com/github/gem-builder.git-4                   	    1000	   1955303 ns/op	  669316 B/op	    1758 allocs/op
BenchmarkDecode/https://github.com/githubtraining/example-branches.git-4      	    3000	    505693 ns/op	  154663 B/op	     545 allocs/op
BenchmarkDecode/https://github.com/rumpkernel/rumprun-xen.git-4               	      10	 189519715 ns/op	94689960 B/op	   67404 allocs/op
BenchmarkDecode/https://github.com/mcuadros/skeetr.git-4                      	     300	   5271772 ns/op	 1265433 B/op	    5437 allocs/op
BenchmarkDecode/https://github.com/dezfowler/LiteMock.git-4                   	     200	   6750296 ns/op	 3134396 B/op	    1238 allocs/op
BenchmarkDecode/https://github.com/tyba/storable.git-4                        	      50	  25293456 ns/op	14527944 B/op	   24515 allocs/op
BenchmarkDecode/https://github.com/toqueteos/ts3.git-4                        	     500	   2788570 ns/op	  633714 B/op	    2402 allocs/op
PASS
ok  	gopkg.in/src-d/go-git.v4/plumbing/format/packfile	28.533s
```

Now, using `Parser`:

```
go test -benchmem -run=^$ gopkg.in/src-d/go-git.v4/plumbing/format/packfile -bench ^BenchmarkParse$

goos: darwin
goarch: amd64
pkg: gopkg.in/src-d/go-git.v4/plumbing/format/packfile
BenchmarkParse/https://github.com/git-fixtures/root-references.git-4         	     300	   5304848 ns/op	 1575867 B/op	    1737 allocs/op
BenchmarkParse/https://github.com/git-fixtures/basic.git-4                   	     300	   4469931 ns/op	 1519526 B/op	     925 allocs/op
BenchmarkParse/https://github.com/git-fixtures/basic.git#01-4                	     300	   4508625 ns/op	 1517858 B/op	     897 allocs/op
BenchmarkParse/https://github.com/git-fixtures/basic.git#02-4                	     300	   4321764 ns/op	 1506340 B/op	     786 allocs/op
BenchmarkParse/https://github.com/src-d/go-git.git-4                         	       2	 912509994 ns/op	146944992 B/op	  110160 allocs/op
BenchmarkParse/https://github.com/git-fixtures/tags.git-4                    	    5000	    275472 ns/op	  121125 B/op	     200 allocs/op
BenchmarkParse/https://github.com/spinnaker/spinnaker.git-4                  	       5	 275443394 ns/op	50149379 B/op	  167296 allocs/op
BenchmarkParse/https://github.com/jamesob/desk.git-4                         	      50	  38003653 ns/op	 6258682 B/op	   19182 allocs/op
BenchmarkParse/https://github.com/cpcs499/Final_Pres_P.git-4                 	   10000	    103859 ns/op	   54202 B/op	      71 allocs/op
BenchmarkParse/https://github.com/github/gem-builder.git-4                   	     500	   3510747 ns/op	  635228 B/op	    2817 allocs/op
BenchmarkParse/https://github.com/githubtraining/example-branches.git-4      	    2000	    945524 ns/op	  163489 B/op	     771 allocs/op
BenchmarkParse/https://github.com/rumpkernel/rumprun-xen.git-4               	       5	 270819780 ns/op	63493952 B/op	  108893 allocs/op
BenchmarkParse/https://github.com/mcuadros/skeetr.git-4                      	     100	  10074921 ns/op	 1290210 B/op	    8008 allocs/op
BenchmarkParse/https://github.com/dezfowler/LiteMock.git-4                   	     100	  12084934 ns/op	 5119299 B/op	    1780 allocs/op
BenchmarkParse/https://github.com/tyba/storable.git-4                        	      30	  50910581 ns/op	10763067 B/op	   40452 allocs/op
BenchmarkParse/https://github.com/toqueteos/ts3.git-4                        	     300	   5518369 ns/op	  571485 B/op	    3662 allocs/op
PASS
ok  	gopkg.in/src-d/go-git.v4/plumbing/format/packfile	29.596s
```

As we can see, Parser is slower. Like, way slower.