@erizocosmico 2018-08-09
Added benchmarks for PackfileIter.

Before:

```
go test -benchmem -run='^$' gopkg.in/src-d/go-git.v4/storage/filesystem -bench '^BenchmarkPackfileIter$'  -v
goos: darwin
goarch: amd64
pkg: gopkg.in/src-d/go-git.v4/storage/filesystem
BenchmarkPackfileIter/https://github.com/git-fixtures/root-references.git-4                  2009890771 ns/op    1417207 B/op       5455 allocs/op
BenchmarkPackfileIter/https://github.com/git-fixtures/basic.git-4                            2008537152 ns/op    1303230 B/op       3957 allocs/op
BenchmarkPackfileIter/https://github.com/git-fixtures/basic.git#01-4                         2008423076 ns/op    1282282 B/op       3905 allocs/op
BenchmarkPackfileIter/https://github.com/git-fixtures/basic.git#02-4                         2008468932 ns/op    1269857 B/op       3715 allocs/op
BenchmarkPackfileIter/https://github.com/git-fixtures/basic.git#03-4                         2008487332 ns/op    1293435 B/op       3898 allocs/op
BenchmarkPackfileIter/https://github.com/git-fixtures/basic.git#04-4                         2009250733 ns/op    1293650 B/op       3898 allocs/op
BenchmarkPackfileIter/https://github.com/git-fixtures/basic.git#05-4                         100          10876467 ns/op    1303022 B/op       3915 allocs/op
BenchmarkPackfileIter/https://github.com/git-fixtures/basic.git#06-4                         2009743280 ns/op    1293033 B/op       3897 allocs/op
BenchmarkPackfileIter/https://github.com/src-d/go-git.git-4                                    1        1948470823 ns/op   216805032 B/op    204000 allocs/op
BenchmarkPackfileIter/https://github.com/git-fixtures/tags.git-4                            2000 972080 ns/op     293142 B/op       2735 allocs/op
BenchmarkPackfileIter/https://github.com/git-fixtures/empty.git-4                          10000 105937 ns/op      25267 B/op         49 allocs/op
PASS
ok      gopkg.in/src-d/go-git.v4/storage/filesystem     24.368s
```

Now:

```
go test -benchmem -run='^$' gopkg.in/src-d/go-git.v4/storage/filesystem -bench '^BenchmarkPackfileIter$'  -v
goos: darwin
goarch: amd64
pkg: gopkg.in/src-d/go-git.v4/storage/filesystem
BenchmarkPackfileIter/https://github.com/git-fixtures/root-references.git-4                  3005015342 ns/op     568647 B/op       7988 allocs/op
BenchmarkPackfileIter/https://github.com/git-fixtures/basic.git-4                            5002899767 ns/op     420181 B/op       5079 allocs/op
BenchmarkPackfileIter/https://github.com/git-fixtures/basic.git#01-4                         5002885262 ns/op     426862 B/op       5256 allocs/op
BenchmarkPackfileIter/https://github.com/git-fixtures/basic.git#02-4                         5002445589 ns/op     407056 B/op       4765 allocs/op
BenchmarkPackfileIter/https://github.com/git-fixtures/basic.git#03-4                         5003050457 ns/op     421045 B/op       5071 allocs/op
BenchmarkPackfileIter/https://github.com/git-fixtures/basic.git#04-4                         5002930551 ns/op     421278 B/op       5070 allocs/op
BenchmarkPackfileIter/https://github.com/git-fixtures/basic.git#05-4                         5002949758 ns/op     421219 B/op       5070 allocs/op
BenchmarkPackfileIter/https://github.com/git-fixtures/basic.git#06-4                         5002936594 ns/op     421052 B/op       5071 allocs/op
BenchmarkPackfileIter/https://github.com/src-d/go-git.git-4                                   10         164301716 ns/op    9509232 B/op     183202 allocs/op
BenchmarkPackfileIter/https://github.com/git-fixtures/tags.git-4                            2000 986470 ns/op     313573 B/op       3010 allocs/op
BenchmarkPackfileIter/https://github.com/git-fixtures/empty.git-4                          20000  80880 ns/op      25241 B/op         49 allocs/op
PASS
ok      gopkg.in/src-d/go-git.v4/storage/filesystem     23.787s
```